requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
idna==2.6
meetup-api==0.1.1
pkg-resources==0.0.0
PyYAML==3.12
requests==2.18.4
six==1.10.0
urllib3==1.22
